+++ src/patpy/pl/__init__.py
@@ -1,1 +1,1 @@
-from .basic import correlation_volcano
+from .basic import correlation_volcano, embedding_covariate_heatmap